Refactor and split out into folders, fixing some lints along the way
